backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
pydantic==1.10.13
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
httpx==0.25.2
